feat: remove unused authentication and middleware extensions, consolidate service registrations in Program.cs
