--- backend/API/Extensions/AuthenticationExtensions.cs
@@ -1,64 +1,0 @@
-﻿using System.Text;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authentication.Cookies;
-using Microsoft.AspNetCore.Authentication.Google;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.IdentityModel.Tokens;
-
-namespace API.Extensions;
-
-public static class AuthenticationExtensions
-{
-    public static IServiceCollection AddJwtAuthentication(this IServiceCollection services)
-    {
-        var jwtKey = Environment.GetEnvironmentVariable("JWT_KEY") ?? "JWT_KEY is missing";
-        var jwtIssuer = Environment.GetEnvironmentVariable("JWT_ISSUER") ?? "default_issuer";
-        var jwtAudience = Environment.GetEnvironmentVariable("JWT_AUDIENCE") ?? "default_audience";
-
-        services.AddAuthentication(options =>
-        {
-            options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-            options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-        })
-        .AddJwtBearer(options =>
-        {
-            options.TokenValidationParameters = new TokenValidationParameters
-            {
-                ValidateIssuer = true,
-                ValidateAudience = true,
-                ValidateLifetime = true,
-                ValidateIssuerSigningKey = true,
-                ValidIssuer = jwtIssuer,
-                ValidAudience = jwtAudience,
-                IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(jwtKey))
-            };
-        });
-
-        return services;
-    }
-
-    public static IServiceCollection AddGoogleOAuth2(this IServiceCollection services)
-    {
-        var googleClientId = Environment.GetEnvironmentVariable("GOOGLE_CLIENT_ID")
-                             ?? throw new InvalidOperationException("Google Client ID is missing.");
-        var googleClientSecret = Environment.GetEnvironmentVariable("GOOGLE_CLIENT_SECRET")
-                                 ?? throw new InvalidOperationException("Google Client Secret is missing.");
-
-        services.AddAuthentication(options =>
-        {
-            options.DefaultAuthenticateScheme = CookieAuthenticationDefaults.AuthenticationScheme;
-            options.DefaultChallengeScheme = GoogleDefaults.AuthenticationScheme;
-            options.DefaultSignInScheme = CookieAuthenticationDefaults.AuthenticationScheme;
-        })
-        .AddCookie()
-        .AddGoogle(options =>
-        {
-            options.ClientId = googleClientId;
-            options.ClientSecret = googleClientSecret;
-            options.SaveTokens = true;
-            options.ClaimActions.MapJsonKey("picture", "picture", "url");
-        });
-
-        return services;
-    }
-}
