--- backend/API/Extensions/SwaggerExtensions.cs
@@ -1,39 +1,0 @@
-﻿using Microsoft.OpenApi.Models;
-
-namespace API.Extensions;
-
-public static class SwaggerExtensions
-{
-    public static IServiceCollection AddSwaggerDocumentation(this IServiceCollection services)
-    {
-        services.AddSwaggerGen(options =>
-        {
-            options.SwaggerDoc("v1", new OpenApiInfo { Title = "GenCare", Version = "v1" });
-            options.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
-            {
-                Name = "Authorization",
-                Type = SecuritySchemeType.Http,
-                Scheme = "Bearer",
-                BearerFormat = "JWT",
-                In = ParameterLocation.Header,
-                Description = "Nhập token theo dạng: Bearer {your token}"
-            });
-            options.AddSecurityRequirement(new OpenApiSecurityRequirement
-            {
-                    {
-                        new OpenApiSecurityScheme
-                        {
-                            Reference = new OpenApiReference
-                            {
-                                Type = ReferenceType.SecurityScheme,
-                                Id = "Bearer"
-                            }
-                        },
-                        Array.Empty<string>()
-                    }
-            });
-        });
-
-        return services;
-    }
-}
